Remove componente filteresList  - Com a alteração no backend do filtro enviar string vazia trás a consulta completa.
